feat: fix big num compile.
