+++ bridge/bindings/qjs/qjs_patch.h
@@ -8,92 +8,7 @@
 
 #include <quickjs/list.h>
 #include <quickjs/quickjs.h>
-
-struct JSString {
-  JSRefCountHeader header; /* must come first, 32-bit */
-  uint32_t len : 31;
-  uint8_t is_wide_char : 1; /* 0 = 8 bits, 1 = 16 bits characters */
-  /* for JS_ATOM_TYPE_SYMBOL: hash = 0, atom_type = 3,
-     for JS_ATOM_TYPE_PRIVATE: hash = 1, atom_type = 3
-     XXX: could change encoding to have one more bit in hash */
-  uint32_t hash : 30;
-  uint8_t atom_type : 2; /* != 0 if atom, JS_ATOM_TYPE_x */
-  uint32_t hash_next;    /* atom_index for JS_ATOM_TYPE_SYMBOL */
-#ifdef DUMP_LEAKS
-  struct list_head link; /* string list */
-#endif
-  union {
-    uint8_t str8[0]; /* 8 bit strings will get an extra null terminator */
-    uint16_t str16[0];
-  } u;
-};
-
-enum {
-  /* classid tag        */ /* union usage   | properties */
-  JS_CLASS_OBJECT = 1,     /* must be first */
-  JS_CLASS_ARRAY,          /* u.array       | length */
-  JS_CLASS_ERROR,
-  JS_CLASS_NUMBER,           /* u.object_data */
-  JS_CLASS_STRING,           /* u.object_data */
-  JS_CLASS_BOOLEAN,          /* u.object_data */
-  JS_CLASS_SYMBOL,           /* u.object_data */
-  JS_CLASS_ARGUMENTS,        /* u.array       | length */
-  JS_CLASS_MAPPED_ARGUMENTS, /*               | length */
-  JS_CLASS_DATE,             /* u.object_data */
-  JS_CLASS_MODULE_NS,
-  JS_CLASS_C_FUNCTION,          /* u.cfunc */
-  JS_CLASS_BYTECODE_FUNCTION,   /* u.func */
-  JS_CLASS_BOUND_FUNCTION,      /* u.bound_function */
-  JS_CLASS_C_FUNCTION_DATA,     /* u.c_function_data_record */
-  JS_CLASS_GENERATOR_FUNCTION,  /* u.func */
-  JS_CLASS_FOR_IN_ITERATOR,     /* u.for_in_iterator */
-  JS_CLASS_REGEXP,              /* u.regexp */
-  JS_CLASS_ARRAY_BUFFER,        /* u.array_buffer */
-  JS_CLASS_SHARED_ARRAY_BUFFER, /* u.array_buffer */
-  JS_CLASS_UINT8C_ARRAY,        /* u.array (typed_array) */
-  JS_CLASS_INT8_ARRAY,          /* u.array (typed_array) */
-  JS_CLASS_UINT8_ARRAY,         /* u.array (typed_array) */
-  JS_CLASS_INT16_ARRAY,         /* u.array (typed_array) */
-  JS_CLASS_UINT16_ARRAY,        /* u.array (typed_array) */
-  JS_CLASS_INT32_ARRAY,         /* u.array (typed_array) */
-  JS_CLASS_UINT32_ARRAY,        /* u.array (typed_array) */
-#ifdef CONFIG_BIGNUM
-  JS_CLASS_BIG_INT64_ARRAY,  /* u.array (typed_array) */
-  JS_CLASS_BIG_UINT64_ARRAY, /* u.array (typed_array) */
-#endif
-  JS_CLASS_FLOAT32_ARRAY, /* u.array (typed_array) */
-  JS_CLASS_FLOAT64_ARRAY, /* u.array (typed_array) */
-  JS_CLASS_DATAVIEW,      /* u.typed_array */
-#ifdef CONFIG_BIGNUM
-  JS_CLASS_BIG_INT,      /* u.object_data */
-  JS_CLASS_BIG_FLOAT,    /* u.object_data */
-  JS_CLASS_FLOAT_ENV,    /* u.float_env */
-  JS_CLASS_BIG_DECIMAL,  /* u.object_data */
-  JS_CLASS_OPERATOR_SET, /* u.operator_set */
-#endif
-  JS_CLASS_MAP,                      /* u.map_state */
-  JS_CLASS_SET,                      /* u.map_state */
-  JS_CLASS_WEAKMAP,                  /* u.map_state */
-  JS_CLASS_WEAKSET,                  /* u.map_state */
-  JS_CLASS_MAP_ITERATOR,             /* u.map_iterator_data */
-  JS_CLASS_SET_ITERATOR,             /* u.map_iterator_data */
-  JS_CLASS_ARRAY_ITERATOR,           /* u.array_iterator_data */
-  JS_CLASS_STRING_ITERATOR,          /* u.array_iterator_data */
-  JS_CLASS_REGEXP_STRING_ITERATOR,   /* u.regexp_string_iterator_data */
-  JS_CLASS_GENERATOR,                /* u.generator_data */
-  JS_CLASS_PROXY,                    /* u.proxy_data */
-  JS_CLASS_PROMISE,                  /* u.promise_data */
-  JS_CLASS_PROMISE_RESOLVE_FUNCTION, /* u.promise_function_data */
-  JS_CLASS_PROMISE_REJECT_FUNCTION,  /* u.promise_function_data */
-  JS_CLASS_ASYNC_FUNCTION,           /* u.func */
-  JS_CLASS_ASYNC_FUNCTION_RESOLVE,   /* u.async_function_data */
-  JS_CLASS_ASYNC_FUNCTION_REJECT,    /* u.async_function_data */
-  JS_CLASS_ASYNC_FROM_SYNC_ITERATOR, /* u.async_from_sync_iterator_data */
-  JS_CLASS_ASYNC_GENERATOR_FUNCTION, /* u.func */
-  JS_CLASS_ASYNC_GENERATOR,          /* u.async_generator_data */
-
-  JS_CLASS_INIT_COUNT, /* last entry for predefined classes */
-};
+#include "quickjs/cutils.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -102,8 +17,8 @@ extern "C" {
 uint16_t* JS_ToUnicode(JSContext* ctx, JSValueConst value, uint32_t* length);
 JSValue JS_NewUnicodeString(JSRuntime* runtime, JSContext* ctx, const uint16_t* code, uint32_t length);
 JSClassID JSValueGetClassId(JSValue);
-bool JS_IsProxy(JSValue value);
-bool JS_HasClassId(JSRuntime* runtime, JSClassID classId);
+BOOL JS_IsProxy(JSValue value);
+BOOL JS_HasClassId(JSRuntime* runtime, JSClassID classId);
 JSValue JS_GetProxyTarget(JSValue value);
 
 #ifdef __cplusplus
